pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factchecker"
version = "0.0.1"
description = "FactChecker: Social media claim verification system"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Chaitanya Rajguru", email = "crajguru@outlook.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "fastapi>=0.104.1",
    "pydantic>=2.0",
    "aiohttp>=3.8.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "tweepy>=4.14.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.5",
    "mypy>=1.7.0",
    "black>=23.12.0",
]

[project.urls]
Homepage = "https://github.com/chaihub/fact-checker"
Documentation = "https://fact-checker.readthedocs.io"
Repository = "https://github.com/chaihub/fact-checker.git"
"Bug Tracker" = "https://github.com/chaihub/fact-checker/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["factchecker"]

[tool.setuptools.package-data]
factchecker = ["py.typed"]

# Ruff configuration
[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "C901",  # McCabe complexity
]

[tool.ruff.isort]
known-first-party = ["factchecker"]
known-third-party = ["fastapi", "pydantic"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "pytesseract",
    "tweepy",
    "aiohttp",
    "google.generativeai",
]
ignore_missing_imports = true

# Black configuration
[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Coverage configuration
[tool.coverage.run]
source = ["src/factchecker"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 85
precision = 2

# LLM Tool Configuration
[tool.llm]
default_provider = "google-gemini"
api_timeout_seconds = 30
max_tokens = 2048
temperature = 0.7
retry_max_attempts = 3
retry_backoff_seconds = 1

[[tool.llm.use_cases]]
name = "claim_extraction_from_text"
description = "Extract structured claims and assertions from user-provided text"
default_model = "gemini-pro"
system_prompt = "Extract structured claims and key assertions from the provided text. Return as JSON with fields: claims (list), key_entities (list), confidence (float 0-1)"

[[tool.llm.use_cases]]
name = "claim_extraction_from_image"
description = "Extract and understand claims from images with OCR text and metadata"
default_model = "gemini-pro-vision"
system_prompt = "Analyze the provided image and extracted text. Generate clarifying questions about who, what, when, where, why. Return as JSON with fields: questions (list), summary (str), confidence (float 0-1)"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/factchecker", "tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "asyncio: marks tests as async",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance tests",
]
